requirements.txt
ultralytics==8.0.227
streamlit==1.29.0
opencv-python==4.8.1.78
pillow==9.5.0
torch==2.1.2
torchvision==0.16.2
pyyaml==6.0.1 